refactor iostreams to cmdutil module
